requirements.txt
azure-cognitiveservices-speech
pyperclip
numpy
sounddevice
python-dotenv
rich
playsound